% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawQQplot.R
\name{drawQQplot}
\alias{drawQQplot}
\title{generate qqplot}
\usage{
drawQQplot(pvector, main = "title", ...)
}
\arguments{
\item{pvector}{(numeric) pvalue vector}

\item{main}{(char) title}

\item{...}{for plotting}
}
\description{
generate qqplot
}
\details{
from http://www.gettinggeneticsdone.com/2010/07/qq-plots-of-p-values-in-r-using-base.html
}
\examples{
drawQQplot(runif(2000),main="random")

# this one has signal
set.seed(42)
pvalues=runif(10000)
pvalues[sample(10000,10)]=pvalues[sample(10000,10)]/5000
drawQQplot(pvalues,"has signal")

}
