% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContour.R
\name{gradient_contour}
\alias{gradient_contour}
\title{plot contour map for 2D data}
\usage{
gradient_contour(df, ttl = "", type = "fill",
  gradLow = "lightgoldenrodyellow", gradHigh = "mediumblue",
  contourBins = 30L, addPoints = TRUE, downSample = 1,
  pt_colour = "black", pt_alpha = 1, addFit = TRUE, fitCol = "maroon",
  ebCol = "mistyrose3", xlab = "x", ylab = "y", xl = NULL, yl = NULL,
  addZeroRef = TRUE, getObj = FALSE, logLog = FALSE)
}
\arguments{
\item{df}{(data.frame) with 'x' and 'y' values}

\item{gradLow, }{gradHigh (char) gradient low and high colour values}

\item{contourBins}{(integer) number of bins in contour}

\item{addPoints}{(logical) if T, shows individual observations}

\item{pt_colour, }{pt_alpha (character) colour for data points; only 
useful if addPoints=TRUE}

\item{addFit}{(logical) add line fit with CI}

\item{fitCol}{(char) colour for line fit}

\item{ebCol}{(char) colour for errorbar}

\item{xlab, }{ylab (char) x/y axis labels}

\item{xl, }{yl (numeric length 2) xlim and ylim. If NULL, shows [1,99]th 
percentile}

\item{addZeroRef}{(logical) add zero as a reference line}

\item{getObj}{(logical) if TRUE returns ggplot object.}

\item{logLog}{(logical) if TRUE plots log-log scale}

\item{fill}{(char) "lines" (contour lines) or "fill" (contour fill)}

\item{downSampls}{(numeric between 0-1)Downsamples data points before
plotting (e.g. 0.1 shows 10% of points)}
}
\description{
plot contour map for 2D data
}
