% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendro.R
\name{plotDendro}
\alias{plotDendro}
\title{plot dendrogram}
\usage{
plotDendro(m, distFun = function(x) {     1 - cor(x) },
  linkageMethod = "average", circ = FALSE, ...)
}
\arguments{
\item{m}{(matrix)}

\item{distFun}{(function) if NULL, uses}

\item{linkageMethod}{(char) for hclust}

\item{circ}{(logical) make circular}

\item{...}{for plot() method}
}
\description{
plot dendrogram
}
