% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendro_clr.R
\name{plotDendro_clr}
\alias{plotDendro_clr}
\title{dendrogram from unsupervised clustering}
\usage{
plotDendro_clr(M, pheno, groupPal, topVar = 10 * 1000L, verbose = TRUE,
  ttl = "", ...)
}
\arguments{
\item{M}{(matrix) data values}

\item{pheno}{(data.frame) row order should match column order of M.
should contain column names for groupBy. Note that groupBy columns 
should be factors}

\item{groupPal}{(list) keys are columns in pheno by which data are to
be grouped. Values could be RColorBrewer palette names (char length 1)
or a vector of colours to use as the palette. In this case
the palette will follow the order of the variable levels}

\item{topVar}{(integer) top most-variable probes to show}

\item{verbose}{(logical) print messages}

\item{...}{parameters for squash::dendromat}
}
\value{
No value. Side effect of printing a dendrogram)
}
\description{
dendrogram from unsupervised clustering
}
