% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing_bar.R
\name{plotMissing_bar}
\alias{plotMissing_bar}
\title{show % samples with missing data, with one bar per variable}
\usage{
plotMissing_bar(x, ttl, showPct = TRUE, tpose = FALSE)
}
\arguments{
\item{x}{(data.frame) rows are variables, columns are samples}

\item{ttl}{(char) plot title}

\item{showPct}{(logical) if TRUE shows value from 0-100 (fraction samples
missing each var). If FALSE shows sample count}

\item{tpose}{(logical) if TRUE transposes matrix, so that % missing samples
is shown}
}
\value{
ggplot barplot showing % missing for each row (variable)
}
\description{
show % samples with missing data, with one bar per variable
}
