% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissMat}
\alias{plotMissMat}
\title{show data missingness as a chequered matrix}
\usage{
plotMissMat(x, outFile = "miss.pdf", wd = 8, ht = 8, xlab = "x",
  ylab = "y", border = "grey50")
}
\arguments{
\item{x}{(matrix) data matrix.}

\item{outFile}{(char) path to file for printing graph}

\item{wd}{(numeric) width in inches}

\item{ht}{(numeric) height in inches}
}
\value{
plots missingness matrix to file
}
\description{
show data missingness as a chequered matrix
}
