% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariables_boxplot.R
\name{plotVariables_boxplot}
\alias{plotVariables_boxplot}
\title{plot unscaled and Z-scaled variable values as boxplot}
\usage{
plotVariables_boxplot(x, cex.x = 8)
}
\arguments{
\item{x}{(data.frame or matrix) rows are variables, columns are samples}

\item{cex.x}{(numeric) font size for x-axis text}
}
\value{
(list) length 2, with names "unscaled" and "scaled". Contains
ggplot objects for the unscaled and scaled boxplot sets respectively.
}
\description{
plot unscaled and Z-scaled variable values as boxplot
}
