% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPropCircles.R
\name{drawPropCircles}
\alias{drawPropCircles}
\title{draw circle with area proportional to number}
\usage{
drawPropCircles(num, sc = 1, ylim = c(-1, 1), texty = 0.5, ...)
}
\arguments{
\item{num}{(numeric) vector of numbers. currently doesn't put labels}

\item{sc}{(numeric) scaling factor}

\item{ylim}{(numeric) plot extent}

\item{texty}{(numeric) position of label above circle}

\item{...}{params for plotrix::draw.circle()}
}
\value{
plots circles with area proportional to number
}
\description{
draw circle with area proportional to number
}
