% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaColored.R
\name{pcaColored}
\alias{pcaColored}
\title{PCA with samples coloured by group properties}
\usage{
pcaColored(m, mdata, groupToPlot, cex = 0.5, ...)
}
\arguments{
\item{m}{(matrix)}

\item{mdata}{(data.frame) metadata. rownames(mdata)=colnames(m)}

\item{groupToPlot}{(char) vector of metadata variables
using which to colour-code samples}
}
\value{
No value. side effect of plotting pca graph
}
\description{
PCA with samples coloured by group properties
}

